Move storage logic to own class
This prevents cases where we need mutexes to guard different sets of
information.
